+++ src/components/eval-bar/eval-bar-display.tsx
