+++ src/components/eval-bar/index.ts
@@ -1,0 +1,3 @@
+import EvalBar from './eval-bar'
+
+export default EvalBar
