+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import ErrorBoundary from './error-boundary'
-import EvalBar from './eval-bar/eval-bar.tsx'
+import ErrorBoundary from './components/error-boundary'
+import EvalBar from './components/eval-bar'
 import './index.css'
 
 createRoot(document.getElementById('root')!).render(
